Add structured logging of all previously-swallowed errors
This change pulls in the log facade with its key-value feature,
and then uses `inspect_err` before each call to `map_err` (and
`is_ok`) to get the errors surfaced, should a user want to see
them.

I chose `warn!` for most of the errors, and `debug!` primarily
for construction/generation tasks, where failures may be less
surprising. I'm not committed to any of the levels, and happy
to remove any that you think aren't useful.